+++ docs/components/plugins/api/APISectionUtils.tsx
@@ -4,6 +4,8 @@ import React from 'react';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 
+import { APIDataType } from './APIDataType';
+
 import { Code, InlineCode } from '~/components/base/code';
 import { H4 } from '~/components/base/headings';
 import Link from '~/components/base/link';
@@ -21,6 +23,7 @@ import { APISectionPlatformTags } from '~/components/plugins/api/APISectionPlatf
 import * as Constants from '~/constants/theme';
 import { Cell, HeaderCell, Row, Table, TableHead } from '~/ui/components/Table';
 import { tableWrapperStyle } from '~/ui/components/Table/Table';
+import { A } from '~/ui/components/Text';
 
 const isDev = process.env.NODE_ENV === 'development';
 
@@ -33,6 +36,7 @@ export enum TypeDocKind {
   Interface = 256,
   Property = 1024,
   Method = 2048,
+  Parameter = 32768,
   TypeAlias = 4194304,
 }
 
@@ -89,9 +93,6 @@ const nonLinkableTypes = [
   'ComponentClass',
   'E',
   'EventSubscription',
-  'File',
-  'FileList',
-  'Manifest',
   'NativeSyntheticEvent',
   'ParsedQs',
   'ServiceActionResult',
@@ -132,7 +133,10 @@ const hardcodedTypeLinks: Record<string, string> = {
   Date: 'https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date',
   Element: 'https://www.typescriptlang.org/docs/handbook/jsx.html#function-component',
   Error: 'https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error',
-  ExpoConfig: 'https://github.com/expo/expo-cli/blob/main/packages/config-types/src/ExpoConfig.ts',
+  ExpoConfig:
+    'https://github.com/expo/expo/blob/main/packages/%40expo/config-types/src/ExpoConfig.ts',
+  File: 'https://developer.mozilla.org/en-US/docs/Web/API/File',
+  FileList: 'https://developer.mozilla.org/en-US/docs/Web/API/FileList',
   MessageEvent: 'https://developer.mozilla.org/en-US/docs/Web/API/MessageEvent',
   Omit: 'https://www.typescriptlang.org/docs/handbook/utility-types.html#omittype-keys',
   Pick: 'https://www.typescriptlang.org/docs/handbook/utility-types.html#picktype-keys',
@@ -280,15 +284,17 @@ export const resolveTypeName = (
     } else if (type === 'reflection' && declaration?.children) {
       return (
         <>
-          {'{ '}
+          {'{\n'}
           {declaration?.children.map((child: PropData, i) => (
             <span key={`reflection-${name}-${i}`}>
+              {'  '}
               {child.name + ': '}
               {resolveTypeName(child.type)}
               {i + 1 !== declaration?.children?.length ? ', ' : null}
+              {'\n'}
             </span>
           ))}
-          {' }'}
+          {'}'}
         </>
       );
     } else if (type === 'tuple' && elements) {
@@ -350,7 +356,7 @@ export const renderParamRow = ({
         {renderFlags(flags, initValue)}
       </Cell>
       <Cell>
-        <InlineCode>{resolveTypeName(type)}</InlineCode>
+        <APIDataType typeDefinition={type} />
       </Cell>
       <Cell>
         <CommentTextBlock
@@ -393,7 +399,8 @@ export const renderDefaultValue = (defaultValue?: string) =>
 
 export const renderTypeOrSignatureType = (
   type?: TypeDefinitionData,
-  signatures?: MethodSignatureData[]
+  signatures?: MethodSignatureData[],
+  allowBlock: boolean = false
 ) => {
   if (signatures && signatures.length) {
     return (
@@ -416,18 +423,35 @@ export const renderTypeOrSignatureType = (
       </InlineCode>
     );
   } else if (type) {
+    if (allowBlock) {
+      return <APIDataType typeDefinition={type} />;
+    }
     return <InlineCode key={`signature-type-${type.name}`}>{resolveTypeName(type)}</InlineCode>;
   }
   return undefined;
 };
 
 export const renderFlags = (flags?: TypePropertyDataFlags, defaultValue?: string) =>
-  flags?.isOptional || defaultValue ? (
+  (flags?.isOptional || defaultValue) && (
     <>
       <br />
       <span css={STYLES_OPTIONAL}>(optional)</span>
     </>
-  ) : undefined;
+  );
+
+export const renderIndexSignature = (kind: TypeDocKind) =>
+  kind === TypeDocKind.Parameter && (
+    <>
+      <br />
+      <A
+        css={STYLES_OPTIONAL}
+        href="https://www.typescriptlang.org/docs/handbook/2/objects.html#index-signatures"
+        openInNewTab
+        isStyled>
+        (index signature)
+      </A>
+    </>
+  );
 
 export type CommentTextBlockProps = {
   comment?: CommentData;
@@ -452,6 +476,13 @@ export const getTagData = (tagName: string, comment?: CommentData) =>
 
 export const getAllTagData = (tagName: string, comment?: CommentData) =>
   comment?.tags?.filter(tag => tag.tag === tagName);
+
+export const getTagNamesList = (comment?: CommentData) =>
+  comment && [
+    ...(getAllTagData('platform', comment)?.map(platformData => platformData.text) || []),
+    ...(getTagData('deprecated', comment) ? ['deprecated'] : []),
+    ...(getTagData('experimental', comment) ? ['experimental'] : []),
+  ];
 
 export const capitalize = (str: string) => str.charAt(0).toUpperCase() + str.slice(1);