+++ packages/expo-av/android/src/main/java/expo/modules/av/player/PlayerData.java
@@ -41,6 +41,7 @@ public abstract class PlayerData implements AudioEventHandler {
   static final String STATUS_RATE_KEY_PATH = "rate";
   static final String STATUS_SHOULD_CORRECT_PITCH_KEY_PATH = "shouldCorrectPitch";
   static final String STATUS_VOLUME_KEY_PATH = "volume";
+  static final String STATUS_VOLUME_PAN_KEY_PATH = "audioPan";
   static final String STATUS_IS_MUTED_KEY_PATH = "isMuted";
   static final String STATUS_IS_LOOPING_KEY_PATH = "isLooping";
   static final String STATUS_DID_JUST_FINISH_KEY_PATH = "didJustFinish";
@@ -102,6 +103,7 @@ public abstract class PlayerData implements AudioEventHandler {
   float mRate = 1.0f;
   boolean mShouldCorrectPitch = false;
   float mVolume = 1.0f;
+  float mPan = 0.0f;
   boolean mIsMuted = false;
 
   PlayerData(final AVManagerInterface avModule, final Uri uri, final Map<String, Object> requestHeaders) {
@@ -324,6 +326,10 @@ public abstract class PlayerData implements AudioEventHandler {
       mVolume = (float) status.getDouble(STATUS_VOLUME_KEY_PATH);
     }
 
+    if (status.containsKey(STATUS_VOLUME_PAN_KEY_PATH)) {
+      mPan = (float) status.getDouble(STATUS_VOLUME_PAN_KEY_PATH);
+    }
+
     if (status.containsKey(STATUS_IS_MUTED_KEY_PATH)) {
       mIsMuted = status.getBoolean(STATUS_IS_MUTED_KEY_PATH);
     }
@@ -420,6 +426,7 @@ public abstract class PlayerData implements AudioEventHandler {
     map.putDouble(STATUS_RATE_KEY_PATH, (double) mRate);
     map.putBoolean(STATUS_SHOULD_CORRECT_PITCH_KEY_PATH, mShouldCorrectPitch);
     map.putDouble(STATUS_VOLUME_KEY_PATH, (double) mVolume);
+    map.putDouble(STATUS_VOLUME_PAN_KEY_PATH, (double) mPan);
     map.putBoolean(STATUS_IS_MUTED_KEY_PATH, mIsMuted);
     // STATUS_IS_LOOPING_KEY_PATH is set in addExtraStatusFields().