+++ packages/expo-modules-core/android/src/main/cpp/JNIInjector.cpp
@@ -4,7 +4,8 @@
 #include "JavaScriptModuleObject.h"
 #include "JavaScriptValue.h"
 #include "JavaScriptObject.h"
-#include "CachedReferencesRegistry.h"
+#include "JavaScriptTypedArray.h"
+#include "JavaReferencesCache.h"
 
 #include <jni.h>
 #include <fbjni/fbjni.h>
@@ -13,11 +14,12 @@
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM *vm, void *) {
   return facebook::jni::initialize(vm, [] {
     // Loads references to often use Java classes
-    expo::CachedReferencesRegistry::instance()->loadJClasses(jni::Environment::current());
+    expo::JavaReferencesCache::instance()->loadJClasses(jni::Environment::current());
 
     expo::JSIInteropModuleRegistry::registerNatives();
     expo::JavaScriptModuleObject::registerNatives();
     expo::JavaScriptValue::registerNatives();
     expo::JavaScriptObject::registerNatives();
+    expo::JavaScriptTypedArray::registerNatives();
   });
 }