+++ packages/expo-modules-core/src/sweet/setUpErrorManager.fx.ts
@@ -1,5 +1,6 @@
 import { EventEmitter } from '../EventEmitter';
 import Platform from '../Platform';
+import { CodedError } from '../errors/CodedError';
 import NativeErrorManager from './NativeErrorManager';
 
 if (__DEV__ && Platform.OS === 'android' && NativeErrorManager) {
@@ -10,3 +11,6 @@ if (__DEV__ && Platform.OS === 'android' && NativeErrorManager) {
     console.error(message);
   });
 }
+
+// We have to export `CodedError` via global object to use in later in the C++ code.
+global.ExpoModulesCore_CodedError = CodedError;