+++ packages/expo-keep-awake/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### 💡 Others
 
+## 10.2.1 — 2022-08-08
+
+### 🐛 Bug fixes
+
+- Attempt to fix `EXC_BAD_ACCESS` and `NSInvalidArgumentException` crashes by not referencing to the class instance function. ([#18553](https://github.com/expo/expo/pull/18553) by [@tsapeta](https://github.com/tsapeta))
+
 ## 10.2.0 — 2022-07-07
 
 ### 🐛 Bug fixes